Uploading Assignment 7, starting assignment 8
Its been a while since my last push, there has been alot going on in my head and alot of work to do recently. ToasterBot needs a major update to work with current JDA on discord, there are alot of projects that need my attention and combine that will all the stuff that is going down at home and I'm finding it hard to get focused. I need to fix the past assignments that aren't completely correct as well.